Refactor the ZipFile class
