+++ common/assemble_versioned/assemble-versioned.py
@@ -25,6 +25,20 @@ import sys
 import zipfile
 import common
 
+# This ZipFile extends Python's ZipFile and fixes the lost permission issue
+class ZipFile(zipfile.ZipFile):
+    def extract(self, member, path=None, pwd=None):
+        if not isinstance(member, zipfile.ZipInfo):
+            member = self.getinfo(member)
+
+        if path is None:
+            path = os.getcwd()
+
+        ret_val = self._extract_member(member, path, pwd)
+        attr = member.external_attr >> 16
+        os.chmod(ret_val, attr)
+        return ret_val
+
 
 output_path = sys.argv[1]
 version_path = sys.argv[2]
